requirements.txt
pandas
numpy
scikit-learn
scipy
fastapi
uvicorn
anthropic
pydantic
python-dotenv
httpx
playwright
